squash: moving the load and loadPatch functions util
* This moves the function that loads the commit, from either a 'git show' or from a url, into the util module
* This is for cleaning up the main cmd.js file which should only really have logic related to the commands
