+++ lib/utils.js
@@ -1,5 +1,9 @@
 'use strict'
 
+const exec = require('child_process').exec
+const http = require('http')
+const https = require('https')
+const url = require('url')
 const chalk = require('chalk')
 const CHECK = chalk.green('✔')
 const X = chalk.red('✖')
@@ -56,4 +60,45 @@ exports.describeSubsystem = function describeSubsystem(subsystems, max = 20) {
       )
     }
   }
+}
+
+// Given a commit hash, load it
+// If a URL is passed in, then load the commit remotely
+// If not, then do a git show
+exports.load = function load(sha, cb) {
+  const parsed = url.parse(sha)
+  if (parsed.protocol) {
+    return exports.loadPatch(parsed, cb)
+  }
+
+  exec(`git show --quiet --format=medium ${sha}`, (err, stdout, stderr) => {
+    if (err) return cb(err)
+    cb(null, stdout.trim())
+  })
+}
+
+// Load the commit from a URL
+exports.loadPatch = function loadPatch(uri, cb) {
+  let h = http
+  if (~uri.protocol.indexOf('https')) {
+    h = https
+  }
+  uri.headers = {
+    'user-agent': 'core-validate-commit'
+  }
+  h.get(uri, (res) => {
+    let buf = ''
+    res.on('data', (chunk) => {
+      buf += chunk
+    })
+
+    res.on('end', () => {
+      try {
+        const out = JSON.parse(buf)
+        cb(null, out)
+      } catch (err) {
+        cb(err)
+      }
+    })
+  }).on('error', cb)
 }
